+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## [Unreleased]
 ### Added
 - Allow selection of multiple ruleset JARs
+- Add github button to settings page
 
 ### Changed